requirements.txt
Django~=4.2.6
django-autoslug~=1.9.9
requests~=2.28.1
django-debug-toolbar~=4.2.0
social-auth-app-django==5.4.0
channels~=4.0.0